+++ src/experimental.ts
@@ -1,0 +1,6 @@
+export type {
+  VkExperimentalCanvasContext,
+  VkExperimentalCanvasTableTransformOptions,
+} from "./experimental-canvas-table-transform.js";
+
+export { createExperimentalCanvasTableTransform } from "./experimental-canvas-table-transform.js";